Added to plot3d library
Working, just need to renumber the faces
